Catch up develop from release

+++ Source/Meadow.Foundation.Peripherals/Displays.Max7219/Driver/Displays.Max7219/Max7219.cs
@@ -9,8 +9,6 @@ namespace Meadow.Foundation.Displays
     /// </summary>
     public class Max7219 : DisplayBase
     {
-        
-
         /// <summary>
         /// MAX7219 Spi Clock Frequency
         /// </summary>
@@ -36,13 +34,9 @@ namespace Meadow.Foundation.Displays
 
         public override DisplayColorMode ColorMode => DisplayColorMode.Format1bpp;
 
-        public override uint Width => (uint)(8 * DeviceColumns);
+        public override int Width => 8 * DeviceColumns;
 
-        public override uint Height => (uint)(8 * DeviceRows);
-
-        
-
-        
+        public override int Height => 8 * DeviceRows;
 
         private ISpiPeripheral max7219;
